GUI2/requirements.txt
groq~=0.9.0
Pillow~=10.4.0
openai~=1.38.0
faster-whisper
pyperclip~=1.9.0
opencv-python~=4.10.0.84
google-generativeai
pyaudio~=0.2.14
SpeechRecognition
customtkinter~=5.2.2
